# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Expense Split Engine API",
    "description": "Enterprise-grade REST API documentation for the Expense Split Engine system.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Base Gateway Endpoint"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "name"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@example.com" },
                  "password": { "type": "string", "minLength": 8, "example": "password123" },
                  "name": { "type": "string", "example": "John Doe" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "email": { "type": "string" },
                        "name": { "type": "string" }
                      }
                    },
                    "token": { "type": "string", "description": "Short-lived access token" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error" },
          "409": { "description": "Email already registered" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Log in and obtain tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@example.com" },
                  "password": { "type": "string", "example": "password123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "headers": {
              "Set-Cookie": {
                "schema": { "type": "string", "example": "refreshToken=xyz...; HttpOnly; Secure; SameSite=Lax" }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "token": { "type": "string", "description": "Short-lived access token" },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "email": { "type": "string" },
                        "name": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Log out user and clear session cookie",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Logout successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": { "type": "string", "example": "Logged out successfully" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Authentication"],
        "summary": "Get currently logged-in user profile",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "User details retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "email": { "type": "string" },
                        "name": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/groups": {
      "post": {
        "tags": ["Groups"],
        "summary": "Create a new group",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "Europe Trip 2026" },
                  "description": { "type": "string", "example": "Shared trip expenses" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "group": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string", "format": "uuid" },
                        "name": { "type": "string" },
                        "description": { "type": "string" }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Groups"],
        "summary": "Retrieve user groups",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Groups list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "groups": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string", "format": "uuid" },
                          "name": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/groups/{groupId}/invites": {
      "post": {
        "tags": ["Groups"],
        "summary": "Invite a user to the group",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "groupId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "invitee@example.com" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invitation sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "inviteCode": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/expenses": {
      "post": {
        "tags": ["Expenses"],
        "summary": "Create a new expense",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "amount", "groupId", "splitType", "participants"],
                "properties": {
                  "title": { "type": "string", "example": "Dinner Party" },
                  "amount": { "type": "integer", "description": "Amount in cents", "example": 5000 },
                  "groupId": { "type": "string", "format": "uuid" },
                  "splitType": { "type": "string", "enum": ["EQUAL", "EXACT", "PERCENTAGE", "SHARE"], "example": "EQUAL" },
                  "participants": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "userId": { "type": "string", "format": "uuid" },
                        "amount": { "type": "integer", "description": "For EXACT splits in cents" },
                        "percentage": { "type": "number", "description": "For PERCENTAGE splits" },
                        "shares": { "type": "integer", "description": "For SHARE splits" }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Expense logged successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "expense": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/ai/categorize-receipt": {
      "post": {
        "tags": ["AI & OCR"],
        "summary": "Use Gemini AI to analyze raw OCR receipt content",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ocrText"],
                "properties": {
                  "ocrText": { "type": "string", "example": "Starbucks Coffee\n12/12/2026\nTotal: 30.00 INR\nCoffee 15.00\nCake 15.00" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "AI categorization and extraction results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "merchant": { "type": "string", "example": "Starbucks" },
                    "amount": { "type": "integer", "description": "In cents", "example": 3000 },
                    "date": { "type": "string", "format": "date", "example": "2026-12-12" },
                    "category": { "type": "string", "example": "FOOD" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/budgets": {
      "post": {
        "tags": ["Budgets"],
        "summary": "Create a budget limit",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["amount", "period", "currency"],
                "properties": {
                  "groupId": { "type": "string", "format": "uuid" },
                  "category": { "type": "string", "example": "FOOD" },
                  "amount": { "type": "integer", "description": "In cents", "example": 25000 },
                  "currency": { "type": "string", "example": "INR" },
                  "period": { "type": "string", "enum": ["MONTHLY", "YEARLY"], "example": "MONTHLY" },
                  "warningThreshold": { "type": "number", "example": 80 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Budget created successfully"
          }
        }
      }
    },
    "/v1/analytics/dashboard": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Retrieve advanced analytics dashboard summary metrics",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Analytics calculations dashboard",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "dashboard": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/status": {
      "get": {
        "tags": ["System & Health"],
        "summary": "Check system status and engine metrics",
        "responses": {
          "200": {
            "description": "Status payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "status": { "type": "string", "example": "UP" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}